+++ src/main/java/dominio/transporte/medios/ParadasTransporte.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class ParadasTransporte {
-    private Integer distanciaAlaSiguiente;
+    private Double distanciaAlaSiguiente;
     private Parada paradaActual;
     private Parada paradaSiguiente;//Cambiar a tipo ParadaTransporte
     private ParadasTransporte paradaTransporteSiguiente;//algo asi
